+++ src/lib/components/hooks/useViewerContext.tsx
@@ -5,6 +5,12 @@ interface IViewerContext extends IViewerOptions {
   setTheme: (theme: 'dark' | 'light') => void
   setShowHelp: (show: boolean) => void
   showHelp: boolean
+  config: {
+    download: boolean
+    share: boolean
+    print: boolean
+    edit: boolean
+  }
 }
 export const ViewerContext = createContext<IViewerContext | null>(null)